Docs: Install GoogleTagManagerMiddleware after the StartSession middleware.
Hopefully this should save other developers a couple of hours of debugging, especially when they upgrade from 5.1. I spend quite a few hours of locating this, while doing said upgrade from 5.1 to 5.4. 

With the Laravel 5.2 using the middleware groups, it's easy to error, and re-order the middlewares. Especially for developers not knowing the order of middleware matters. 

E.g. like when I moved the various middlewares from global to the 'web' middleware, and re-ordered them alphabetically for readability -.- (sigh).